requirements.txt
aiofiles>=23.2.1
crawl4ai>=0.1.0
dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.5.1
usp-tree>=0.1.0
quart>=0.19.0
llm_processor>=0.1.0